more file structure, docs cleanup
