+++ scripts/atest.py
@@ -1,5 +1,4 @@
-""" Run acceptance tests with robot framework
-"""
+"""Run acceptance tests with robot framework"""
 
 # Copyright (c) 2023 ipyforcegraph contributors.
 # Distributed under the terms of the Modified BSD License.
