+++ src/ipyforcegraph/__init__.py
@@ -1,4 +1,5 @@
-"""Main entrypoint for ipyforcegraph."""
+"""Main entrypoint for ``ipyforcegraph``."""
+
 # Copyright (c) 2023 ipyforcegraph contributors.
 # Distributed under the terms of the Modified BSD License.
