+++ src/ipyforcegraph/behaviors/__init__.py
@@ -1,3 +1,5 @@
+"""Behaviors for ``ipyforcegraph`` graphs, nodes, and links."""
+
 # Copyright (c) 2023 ipyforcegraph contributors.
 # Distributed under the terms of the Modified BSD License.
