+++ src/ipyforcegraph/behaviors/graph.py
@@ -1,4 +1,5 @@
 """Graph behaviors for ``ipyforcegraph``."""
+
 # Copyright (c) 2023 ipyforcegraph contributors.
 # Distributed under the terms of the Modified BSD License.
