+++ src/ipyforcegraph/behaviors/node.py
@@ -1,4 +1,5 @@
 """Node behaviors for ``ipyforcegraph``."""
+
 # Copyright (c) 2023 ipyforcegraph contributors.
 # Distributed under the terms of the Modified BSD License.
